+++ CHANGES.md
@@ -1,5 +1,8 @@
 # 0.2.0 (next)
 
+- Added convenience constant `License::CC_ZERO` for the
+  [CC0](https://creativecommons.org/publicdomain/zero/1.0/legalcode) public domain declaration
+- Allow `License` to take a string as a `uri` parameter, so long as that string is a valid URI
 - Validate parameters for:
   - `StashWrapper`
   - `StashAdministrative` (fixes issue #2)